Rewriting downloadtask to use default requests
