+++ build.gradle
@@ -13,9 +13,9 @@ apply plugin: 'maven'
 
 xlDocker {
   compileImage = 'xebialabs/xlr_dev_compile'
-  compileVersion = 'v7.2.0.1'
+  compileVersion = 'v7.5.0.1'
   runImage = 'xebialabs/xlr_dev_run'
-  runVersion = 'v7.2.0.1'
+  runVersion = 'v7.5.0.1'
   runPortMapping = '15516:5516'
 }
