+++ src/main/resources/bitbucket/BitbucketTask.py
@@ -8,9 +8,9 @@
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
 
-from bitbucket.Bitbucket import BitbucketClient
+import bitbucket
 
-bitbucket = BitbucketClient.get_client(server, username, password)
+bitbucket = bitbucket.BitbucketClient.get_client(server, username, password)
 method = str(task.getTaskType()).lower().replace('.', '_')
 call = getattr(bitbucket, method)
 response = call(locals())
